+++ src/ClientFactory/SymfonyFactory.php
@@ -11,10 +11,8 @@ use Symfony\Component\HttpClient\HttplugClient;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
- *
- * @final
  */
-class SymfonyFactory implements ClientFactory
+final class SymfonyFactory implements ClientFactory
 {
     public function __construct(
         private readonly ResponseFactoryInterface $responseFactory,