+++ src/Collector/Collector.php
@@ -18,10 +18,8 @@ use Symfony\Component\HttpKernel\DataCollector\DataCollector;
  * @author Fabien Bourigault <bourigaultfabien@gmail.com>
  *
  * @internal
- *
- * @final
  */
-class Collector extends DataCollector
+final class Collector extends DataCollector
 {
     private ?Stack $activeStack = null;