+++ src/Collector/ProfileClient.php
@@ -21,10 +21,8 @@ use Symfony\Component\Stopwatch\StopwatchEvent;
  * @author Fabien Bourigault <bourigaultfabien@gmail.com>
  *
  * @internal
- *
- * @final
  */
-class ProfileClient implements ClientInterface, HttpAsyncClient
+final class ProfileClient implements ClientInterface, HttpAsyncClient
 {
     use VersionBridgeClient;