+++ src/Collector/StackPlugin.php
@@ -15,10 +15,8 @@ use Psr\Http\Message\ResponseInterface;
  * @author Fabien Bourigault <bourigaultfabien@gmail.com>
  *
  * @internal
- *
- * @final
  */
-class StackPlugin implements Plugin
+final class StackPlugin implements Plugin
 {
     use Plugin\VersionBridgePlugin;