+++ src/Discovery/ConfiguredClientsStrategy.php
@@ -14,10 +14,8 @@ use Psr\Http\Client\ClientInterface;
  * we can use the web debug toolbar for clients found with the discovery.
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
- *
- * @final
  */
-class ConfiguredClientsStrategy implements DiscoveryStrategy
+final class ConfiguredClientsStrategy implements DiscoveryStrategy
 {
     private static ?ClientInterface $client = null;